\name{blockra-package}
\alias{blockRA-package}
\alias{blockra}
\docType{package}
\title{
Tools For Block Rearrangement of Arrays
}
\description{
Provide algorithms for rearrangement of array to derive the dependence structure that minimizes the variance of the portfolio losses over all possible states.}
\details{
\tabular{ll}{
Package: \tab blockra\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-05-22\cr
License: \tab GPL (>= 2) \cr
}
}

\author{
Kris Boudt, Steven Vanduffel, Kristof Verbeken

Maintainer: Kris Boudt <Kris.Boudt@vub.ac.be>

Thanks: A special thanks for additional contributions from Carole Bernard.
}

\keyword{ package }

\examples{
# see users manual
}
