\name{bra}
\alias{bra}
\title{bra}
\description{%
  Function which performs the block ra of Bernard and McLeish}
\usage{%
  bra(X, epsilon=0.1)
}
\arguments{%
  \item{X}{array or matrix}
  \item{epsilon}{convergence criterium}
}
\details{%

}
\value{%
  \code{X}: rearranged matrix.\cr
}
\note{%
 
}
\author{%
  Kris Boudt, Steven Vanduffel, Kristof Verbeken
}
\references{%

}
\seealso{%
  \code{\link{ra}} and \code{\link{brave}}.
}
\keyword{RA}
