\name{BRA2}
\alias{BRA2}
\title{BRA2}
\description{%
  Function which performs the stochastic block ra of Bernard and McLeish}
\usage{%
  BRA2(X, epsilon=0.1)
}
\arguments{%
  \item{X}{array or matrix}
  \item{epsilon}{convergence criterium}
}
\details{%

}
\value{%
  \code{X}: rerranged matrix.\cr
}
\note{%
 
}
\author{%
  Kris Boudt, Steven Vanduffel, Kristof Verbeken
}
\references{%

}
\seealso{%
  \code{\link{RA}} and \code{\link{BRAVE}}.
}
\keyword{RA}
