% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/brave.r
\name{brave}
\alias{brave}
\title{Block ReArrangement Variance Equalizer}
\usage{
brave(matrix, epsilon = 0.1, shuffle = TRUE)
}
\arguments{
\item{matrix}{numeric matrix}

\item{epsilon}{target variance of row sums is epsilon multiplied by the mean of the matrix variances}

\item{shuffle}{randomly permute each column of the matrix before rearrangement}
}
\value{
numeric matrix with a minimal row sum variance
}
\description{
Block ReArrangement Variance Equalizer
}
\examples{
brave(matrix)
brave(matrix(1:5, 5, 3))
brave(matrix, epsilon = 0.001, shuffle = FALSE)
}

