% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/equalvar.r
\name{equalvar}
\alias{equalvar}
\title{Find two blocks with equal variance}
\usage{
equalvar(matrix)
}
\arguments{
\item{matrix}{numeric matrix}
}
\value{
numeric binary vector with the partition
}
\description{
Find two blocks with equal variance
}
\examples{
equalvar(matrix)
equalvar(matrix(1:5, 5, 3))
}

