% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/partitionproblem.r
\name{partition}
\alias{partition}
\title{Inversely sort a block based on the order of a bigger block}
\usage{
partition(x, n = 2, method = ra, epsilon = 0.1)
}
\arguments{
\item{x}{vector with numbers to be parititioned}

\item{n}{number of block to parition the vector in}
}
\value{
vector with indexes of the blocks
}
\description{
Inversely sort a block based on the order of a bigger block
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}

