% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rearrange.r
\name{rearrange}
\alias{rearrange}
\title{Inversely sort a block based on the order of a bigger block}
\usage{
rearrange(block1, block2)
}
\arguments{
\item{block1}{the small block to be inversely rearranged}

\item{block2}{the large block}
}
\value{
rearranged first block
}
\description{
Inversely sort a block based on the order of a bigger block
}
\examples{
rearrange(block1, block2)
}

