% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shufflematrix.r
\name{shufflematrix}
\alias{shufflematrix}
\title{Shuffle matrix}
\usage{
shufflematrix(X)
}
\arguments{
\item{X}{numeric array or matrix}

\item{fix.first}{don't change the order of the first column}
}
\value{
randomly permuted numeric matrix
}
\description{
Permute randomly each column of the matrix, except for the first column
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Edgars Jakobsons, \email{edgars.jakobsons@math.ethz.ch}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\references{
\url{http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2634669}
}
\seealso{
The \code{\link{ra}} for the rearrangement algorithm

\code{\link{ra}} for the rearrangement algorithm

\code{\link{blockra}} for the block rearrangement algorithm
}

