% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bra.r
\name{bra}
\alias{bra}
\title{Block rearrangement algorithm}
\usage{
bra(X, epsilon = 0.1, shuffle = TRUE, fix.first = TRUE, obj = var)
}
\arguments{
\item{X}{numeric array or matrix}

\item{epsilon}{target variance of row sums is epsilon multiplied by the mean of the matrix variances}

\item{shuffle}{randomly permute each column of the matrix before rearrangement}

\item{fix.first}{don't change the order of the first column}

\item{obj}{objective function that is minimized, default is variance}
}
\value{
numeric matrix with a minimal row sum variance
}
\description{
Function which performs the block ra of Bernard and McLeish
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\references{
\url{LINK TO RA PAPER}
}
\seealso{
The \code{\link{ra}} for the rearrangement algorithm
}

