\name{BRAVE}
\alias{BRAVE}
\title{BRAVE}
\description{%
  Function which performs BRAVE}
\usage{%
  BRAVE(X, epsilon=0.1)
}
\arguments{%
  \item{X}{array or matrix}
  \item{epsilon}{convergence criterium}
}
\details{%

}
\value{%
  \code{X}: rerranged matrix.\cr
}
\note{%
 
}
\author{%
  Kris Boudt, Steven Vanduffel, Kristof Verbeken
}
\references{%

}
\seealso{%
  \code{\link{BRA2}} and \code{\link{RA}}.
}
\keyword{RA}
