\name{brave}
\alias{brave}
\title{brave}
\description{%
  Function which performs brave}
\usage{%
  brave(X, epsilon=0.1)
}
\arguments{%
  \item{X}{array or matrix}
  \item{epsilon}{convergence criterium}
}
\details{%

}
\value{%
  \code{X}: rearranged matrix.\cr
}
\note{%
 
}
\author{%
  Kris Boudt, Steven Vanduffel, Kristof Verbeken
}
\references{%

}
\seealso{%
  \code{\link{bra}} and \code{\link{ra}}.
}
\keyword{RA}
