% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brave.r
\name{brave}
\alias{brave}
\title{Block ReArrangement Variance Equalizer}
\usage{
brave(X, f = var, shuffle = F, maxiter = 1000, stalliter = maxiter,
  abs.tol = 0, rel.tol = 0, f.target = -Inf, version = 1,
  nochange.tol = 0, method = "KK")
}
\arguments{
\item{f}{function of the rowsums to be minimized}

\item{shuffle}{randomly permute each column of the matrix before rearrangement}

\item{maxiter}{number of maximum iterations}

\item{stalliter}{convergence if no improvement after stalliter iteration}

\item{abs.tol}{abs convergence crit}

\item{rel.tol}{relative convergence criterion}

\item{f.target}{value of the objective function at which absolute converge is reached}

\item{version}{version of the brave use}

\item{nochange.tol}{switch to random block condition}

\item{method}{partitioning method used}
}
\value{
list of rearranged matrix, rowsums (descending), iterations, current objective, iterations of objective, whether converged
}
\description{
Block ReArrangement Variance Equalizer
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Edgars Jakobsons, \email{edgars.jakobsons@math.ethz.ch}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\references{
\url{http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2634669}
}
\seealso{
\code{\link{ra}} for the rearrangement algorithm

\code{\link{blockra}} for the block rearrangement algorithm
}

