% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/equalvar.r
\name{equalvar}
\alias{equalvar}
\title{Find two blocks with equal variance}
\usage{
equalvar(X)
}
\arguments{
\item{X}{numeric array or matrix}
}
\value{
numeric binary vector with the partition
}
\description{
The matrix is divided in two blocks with an equal variance using the greedy
implementation of the partition problem on the covariances between the columns
and the total row sums.
}
\details{
The objective is to find two block with an variance that is as close as
possible to the variance of the other block without significantly impacting
the performance.

The first step is to calculate the covariances between all columns and
the total row sums of the matrix. Using the greedy implementation of the
partition problem, we find two groups with variances that are as close as
possible to eachother.
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\references{
\url{Lhttp://en.wikipedia.org/wiki/Partition_problem#The_greedy_algorithm}
}
\keyword{internal}

