\name{RA}
\alias{RA}
\title{RA}
\description{%
  Function which performs the traditional ra}
\usage{%
  RA(X, epsilon=0.1)
}
\arguments{%
  \item{X}{array or matrix}
  \item{epsilon}{convergence criterium}
}
\details{%

}
\value{%
  \code{X}: rerranged matrix.\cr
}
\note{%
 
}
\author{%
  Kris Boudt, Steven Vanduffel, Kristof Verbeken
}
\references{%

}
\seealso{%
  \code{\link{BRA2}} and \code{\link{BRAVE}}.
}
\keyword{RA}
