\name{ra}
\alias{ra}
\title{ra}
\description{%
  Function which performs the traditional ra}
\usage{%
  ra(X, epsilon=0.1)
}
\arguments{%
  \item{X}{array or matrix}
  \item{epsilon}{convergence criterium}
}
\details{%

}
\value{%
  \code{X}: rerranged matrix.\cr
}
\note{%
 
}
\author{%
  Kris Boudt, Steven Vanduffel, Kristof Verbeken
}
\references{%

}
\seealso{%
  \code{\link{bra}} and \code{\link{brave}}.
}
\keyword{RA}
