% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.r
\name{rearrange}
\alias{rearrange}
\title{Inversely sort a block based on the order of a bigger block}
\usage{
rearrange(block1, block2)
}
\arguments{
\item{block1}{the small block to be inversely rearranged}

\item{block2}{the large block}
}
\value{
rearranged first block
}
\description{
Inversely sort a block based on the order of a bigger block
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Edgars Jakobsons, \email{edgars.jakobsons@math.ethz.ch}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\references{
\url{http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2634669}
}
\seealso{
\code{\link{ra}} for the rearrangement algorithm

\code{\link{blockra}} for the block rearrangement algorithm
}
\keyword{internal}

