% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangepartition.r
\name{rearrangepartition}
\alias{rearrangepartition}
\title{Rearrange matrix based on binary partition vector}
\usage{
rearrangepartition(X, partition)
}
\arguments{
\item{X}{numeric array or matrix}

\item{partition}{binary partition vector with 1 representing block 1
and 0 representing block 2}

\item{fix.first}{don't change the order of the first column}
}
\value{
numeric matrix with a minimal row sum variance
}
\description{
Rearrange matrix based on binary partition vector
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Edgars Jakobsons, \email{edgars.jakobsons@math.ethz.ch}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\references{
\url{http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2634669}
}
\seealso{
\code{\link{ra}} for the rearrangement algorithm

\code{\link{blockra}} for the block rearrangement algorithm
}

