% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/shufflematrix.r
\name{shufflematrix}
\alias{shufflematrix}
\title{Shuffle matrix}
\usage{
shufflematrix(X)
}
\arguments{
\item{X}{numeric array or matrix}
}
\value{
randomly permuted numeric matrix
}
\description{
Permute randomly each column of the matrix, except for the first column
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\references{
\url{LINK TO RA PAPER}
}
\seealso{
The \code{\link{ra}} for the rearrangement algorithm
}

